쉐어 참여 취소 API 누락된 실패 케이스와 채팅방까지 나가기 구현

+++ src/main/java/louie/hanse/shareplate/service/EntryService.java
@@ -51,8 +51,8 @@ public class EntryService {
 
     @Transactional
     public void cancel(Long shareId, Long memberId) {
-        memberService.findByIdOrElseThrow(memberId);
-        Share share = shareService.findByIdOrElseThrow(shareId);
+        Member member = memberService.findByIdOrElseThrow(memberId);
+        Share share = shareService.findWithWriterByIdOrElseThrow(shareId);
 
         share.isCanceledThrowException();
         if (!isExistEntry(shareId, memberId)) {
@@ -64,7 +64,12 @@ public class EntryService {
         if (share.isLeftLessThanAnHour()) {
             throw new GlobalException(EntryExceptionType.CLOSE_TO_THE_CLOSED_DATE_TIME);
         }
+        if (share.isWriter(member)) {
+            throw new GlobalException(EntryExceptionType.SHARE_WRITER_CANNOT_ENTRY_CANCEL);
+        }
         entryRepository.deleteByMemberIdAndShareId(memberId, shareId);
+        chatRoomMemberRepository.deleteByMemberIdAndChatRoomId(
+            memberId, share.getEntryChatRoom().getId());
     }
 
     public List<Long> getIdList(Long memberId) {
